requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
easyocr
Pillow
pytesseract
